pyproject.toml
[project]
name = "hexlet-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygithub>=2.6.1",
    "pytest>=8.3.5",
]
